pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "crmproject"
version = "0.1.0"
description = "A CRM project built with Python"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=1.10.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0"
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
